Switch default PCCS port number from 8081 to 10801
Port 8081 is a very poor choice of port number, since it is
both assigned to existing softrware in /etc/services, and
a fairly common "alternative" HTTP port that application
developers use for ad-hoc services.

Move it to 10801 which is not assigned in /etc/services and
thus unlikely to clash with other software.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>