Workaround broken GCC 15
The C++ standard does not allow 'alignas' to be applied to type
aliases. GNU C++ has allowed it as a non-standard extension, but
strictly that should have been expressed using 'gnu::aligned'
instead of 'alignas'. Regardless of the syntax, however, GCC 15
has a regression causing it to reject alignment requests entirely
on template aliases.

https://gcc.gnu.org/bugzilla/show_bug.cgi?id=118773
Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>